requirements.txt
streamlit>=1.50.0
requests>=2.31.0
Pillow>=12.0.0
python-dotenv>=1.0.0
stability-sdk>=0.8.6